moved save/select over to app.rs
